+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Luxembourg country rules
+
 ## [4.27.2] - 2025-12-09
 
 ### Fixed